requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
